Conflicts:
	src/server/game/Entities/Player/Player.cpp
	src/server/game/Entities/Player/Player.h
	src/server/game/Entities/Unit/Unit.cpp
	src/server/scripts/Pandaria/MogushanVault/boss_elegon.cpp
	src/server/scripts/Pandaria/MogushanVault/boss_garajal.cpp
	src/server/scripts/Pandaria/MogushanVault/boss_spirit_kings.cpp
	src/server/scripts/Pandaria/MogushanVault/boss_stone_guard.cpp
	src/server/scripts/Pandaria/MogushanVault/boss_will_of_emperor.cpp
	src/server/scripts/Pandaria/MogushanVault/instance_mogu_shan_vault.cpp

+++ src/server/game/Achievements/AchievementMgr.h
@@ -176,7 +176,7 @@ struct AchievementCriteriaData
     }
 
     bool IsValid(AchievementCriteriaEntry const* criteria);
-    bool Meets(uint32 criteria_id, Player const* source, Unit const* target, uint32 miscvalue1 = 0) const;
+    bool Meets(uint32 criteria_id, Player const* source, Unit const* target, uint64 miscvalue1 = 0) const;
 };
 
 struct AchievementCriteriaDataSet
@@ -184,7 +184,7 @@ struct AchievementCriteriaDataSet
         AchievementCriteriaDataSet() : criteria_id(0) {}
         typedef std::vector<AchievementCriteriaData> Storage;
         void Add(AchievementCriteriaData const& data) { storage.push_back(data); }
-        bool Meets(Player const* source, Unit const* target, uint32 miscvalue = 0) const;
+        bool Meets(Player const* source, Unit const* target, uint64 miscvalue = 0) const;
         void SetCriteriaId(uint32 id) {criteria_id = id;}
     private:
         uint32 criteria_id;
